requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
loguru==0.7.2
edge-tts==7.0.0
anyio==4.3.0
starlette==0.37.2
openai==1.25.1
async-timeout==4.0.3
emoji==2.12.1
aiofiles==23.2.1
aiohttp==3.11.10
python-dotenv==1.0.1
pydub==0.25.1